core code updates
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@3359 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/StatementResultUpdate.java
@@ -143,7 +143,8 @@ public class StatementResultUpdate extends StatementDML {
 
         if (rowIdIndex + 2 == result.metaData.getExtendedColumnCount()) {
             Object[] data =
-                ((RowSetNavigatorData) result.getNavigator()).getData(rowId);
+                ((RowSetNavigatorData) result.getNavigator()).getData(
+                    rowId.longValue());
 
             if (data != null) {
                 row = (Row) data[rowIdIndex + 1];
